+++ Modules/Utils/Sources/PocketCastsUtils/Feature Flags/FeatureFlag.swift
@@ -252,7 +252,7 @@ public enum FeatureFlag: String, CaseIterable {
         case .generatedTranscripts:
             true
         case .podcastViewChanges:
-            false
+            true
         case .podcastNewformAppsFlyer:
             false
         }