+++ podcasts/Strings+Generated.swift
@@ -1710,6 +1710,10 @@ internal enum L10n {
   internal static var notificationsOff: String { return L10n.tr("Localizable", "notifications_off") }
   /// Notifications On
   internal static var notificationsOn: String { return L10n.tr("Localizable", "notifications_on") }
+  /// We’ll notify you with new episodes of %1$@
+  internal static func notificationsOnForPodcast(_ p1: Any) -> String {
+    return L10n.tr("Localizable", "notifications_on_for_podcast", String(describing: p1))
+  }
   /// Play Now
   internal static var notificationsPlayNow: String { return L10n.tr("Localizable", "notifications_play_now") }
   /// Now Playing